Remove ssh.conf
This was causing inconsistencies in the global config, depending on if a
normal tmux session or ssh session was started first.

The environment variables that I actually use are now always updated,
and $SHLVL is always reset to 0. It doesn't mean much anyways unless I'm
directly nesting local tmux sessions, which is a Bad Idea™.

I now set the `destroy-unattached` option on the individual ssh
sessions, so detaching will automatically kill those sessions, making
the extra bindings for d and C-d unnecessary. This also makes the
sessions go away if a connection dies.